+++ package/sdk/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.3.2",
-  genVersion: "2.449.0",
-  userAgent: "speakeasy-sdk/typescript 0.3.2 2.449.0 1.0.0 ryan-total-test-act",
+  sdkVersion: "0.3.3",
+  genVersion: "2.460.1",
+  userAgent: "speakeasy-sdk/typescript 0.3.3 2.460.1 1.0.0 ryan-total-test-act",
 } as const;