+++ package/sdk/RELEASES.md
@@ -218,3 +218,13 @@ Based on:
 - [typescript v0.3.2] package/sdk
 ### Releases
 - [NPM v0.3.2] https://www.npmjs.com/package/ryan-total-test-act/v/0.3.2 - package/sdk
+
+## 2024-11-18 19:30:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.440.2 (2.460.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.3] package/sdk
+### Releases
+- [NPM v0.3.3] https://www.npmjs.com/package/ryan-total-test-act/v/0.3.3 - package/sdk